+++ app/src/main/java/com/savvasdalkitsis/gameframe/feature/control/presenter/ControlPresenter.kt
@@ -24,11 +24,14 @@ import com.savvasdalkitsis.gameframe.feature.ip.repository.IpRepository
 import com.savvasdalkitsis.gameframe.feature.navigation.Navigator
 import com.savvasdalkitsis.gameframe.infra.rx.RxTransformers
 import com.savvasdalkitsis.gameframe.base.BasePresenter
+import com.savvasdalkitsis.gameframe.feature.wifi.model.WifiNotEnabledException
+import com.savvasdalkitsis.gameframe.feature.wifi.usecase.WifiUseCase
 import io.reactivex.Completable
 
 class ControlPresenter(private val gameFrameUseCase: GameFrameUseCase,
                        private val ipRepository: IpRepository,
-                       private val navigator: Navigator): BasePresenter<ControlView>() {
+                       private val navigator: Navigator,
+                       private val wifiUseCase: WifiUseCase): BasePresenter<ControlView>() {
 
     fun loadIpAddress() = stream {
         ipRepository.ipAddress
@@ -59,10 +62,19 @@ class ControlPresenter(private val gameFrameUseCase: GameFrameUseCase,
                     .compose(RxTransformers.schedulers())
 
     private fun runCommandAndNotifyView(command: Completable) = stream {
-        runCommand(command).subscribe({ view?.operationSuccess() }, { e -> view?.operationFailure(e) })
+        runCommand(command).subscribe({ view?.operationSuccess() }, { e ->
+            when (e) {
+                is WifiNotEnabledException -> view?.wifiNotEnabledError(e)
+                else -> view?.operationFailure(e)
+            }
+        })
     }
 
     private fun runCommandAndIgnoreResult(command: Completable) = stream {
         runCommand(command).subscribe({ }, { })
+    }
+
+    fun enableWifi() {
+        wifiUseCase.enableWifi()
     }
 }