+++ app/src/main/assets/changelog.md
@@ -15,6 +15,7 @@ body {
 * Moved to Firebase from fabric
 * Navigating to ip setup if missing when uploading to Game Frame
 * Changing the navigation bar color to the same color as the status bar on lollipop and later devices
+* Warn the user when the wifi is off and option to enable
 
 #### v1.1.2
 ##### 04 Nov 2017